+++ src/main/java/com/xiaoleilu/hutool/util/IoUtil.java
@@ -42,7 +42,7 @@ public class IoUtil {
 	 * @return 拷贝的字节数
 	 * @throws IOException
 	 */
-	public static int copy(Reader reader, Writer writer) throws IOException {
+	public static long copy(Reader reader, Writer writer) throws IOException {
 		return copy(reader, writer, DEFAULT_BUFFER_SIZE);
 	}
 	
@@ -54,9 +54,9 @@ public class IoUtil {
 	 * @return 传输的byte数
 	 * @throws IOException
 	 */
-	public static int copy(Reader reader, Writer writer, int bufferSize) throws IOException {
+	public static long copy(Reader reader, Writer writer, int bufferSize) throws IOException {
 		char[] buffer = new char[bufferSize];
-		int count = 0;
+		long count = 0;
 		int readSize;
 		while ((readSize = reader.read(buffer, 0, bufferSize)) >= 0) {
 			writer.write(buffer, 0, readSize);
@@ -74,7 +74,7 @@ public class IoUtil {
 	 * @return 传输的byte数
 	 * @throws IOException
 	 */
-	public static int copy(InputStream in, OutputStream out) throws IOException {
+	public static long copy(InputStream in, OutputStream out) throws IOException {
 		return copy(in, out, DEFAULT_BUFFER_SIZE);
 	}
 	
@@ -86,9 +86,9 @@ public class IoUtil {
 	 * @return 传输的byte数
 	 * @throws IOException
 	 */
-	public static int copy(InputStream in, OutputStream out, int bufferSize) throws IOException {
+	public static long copy(InputStream in, OutputStream out, int bufferSize) throws IOException {
 		byte[] buffer = new byte[bufferSize];
-		int count = 0;
+		long count = 0;
 		for (int n = -1; (n = in.read(buffer)) != -1;) {
 			out.write(buffer, 0, n);
 			count += n;
